Implemented Steve's REST connection methods, modified a few things to make that happen. Removed and ignored some autogenerated stuff that shouldn't be tracked anyway.

+++ code/mobile/smartFridge/src/edu/rit/smartFridge/util/TestConnect.java
@@ -6,7 +6,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.Random;
 
-import android.content.Context;
+import org.apache.http.client.HttpClient;
+
 import edu.rit.smartFridge.model.InventoryItem;
 import edu.rit.smartFridge.model.ShoppingList;
 
@@ -62,11 +63,11 @@ public class TestConnect implements DataConnect
 		return retList;
 	}
 
-	public ShoppingList getList(int listId)
+	public ShoppingList getList(long listId)
 	{
 		ShoppingList temp;
 
-		switch (listId)
+		switch ((int) listId)
 		{
 			case 0:
 				temp = list1;
